experiment/requirements.txt
djangorestframework
django-filter
django
mysqlclient